+++ src/types.ts
@@ -1,4 +1,3 @@
-import type { Document } from 'mongodb'
 import { JSONSchema } from 'mongochangestream'
 
 export interface ImmutableOption {
@@ -10,9 +9,10 @@ export interface ImmutableOption {
 }
 
 export interface SyncOptions {
-  mapper?: (doc: Document) => Document
   schemaName?: string
   tableName?: string
+  /** Dotted path to renamed dotted path */
+  rename?: Record<string, string>
 }
 
 export interface Override extends Record<string, any> {
@@ -24,6 +24,8 @@ export interface Override extends Record<string, any> {
 export interface ConvertOptions {
   omit?: string[]
   overrides?: Override[]
+  /** Dotted path to renamed dotted path */
+  rename?: Record<string, string>
 }
 
 export type Events = 'process' | 'error'